Extract submission creation from /submit into createSubmission
Move the submission-creation logic (saveUser, required-field and
future-timestamp validation, the Parse object, attachment uploads, and
the final save) out of server.js's /submit handler into
src/createSubmission.js, so it can be unit-tested without the HTTP/
multer/coercion glue. The handler keeps request parsing and response
shaping, and passes saveUser and versionNumber in — both come from
server.js's app configuration.

Behavior is unchanged: the module is a verbatim move, including the
process.env.TZ juggling around the future-timestamp check (which only
restores the previous timezone on the error path).
